+++ test/integration/068_partial_parsing_tests/test-files/snapshot.sql
@@ -13,3 +13,17 @@ select * from {{ ref('orders') }}
 
 {% endsnapshot %}
 
+{% snapshot orders2_snapshot %}
+
+{{
+    config(
+      target_schema=schema,
+      strategy='check',
+      unique_key='id',
+      check_cols=['order_date'],
+    )
+}}
+
+select * from {{ ref('orders') }}
+
+{% endsnapshot %}