+++ test/integration/068_partial_parsing_tests/test-files/snapshot2.sql
@@ -1,3 +1,4 @@
+- add a comment
 {% snapshot orders_snapshot %}
 
 {{
@@ -8,7 +9,22 @@
       check_cols=['status'],
     )
 }}
+
 select * from {{ ref('orders') }}
 
 {% endsnapshot %}
 
+{% snapshot orders2_snapshot %}
+
+{{
+    config(
+      target_schema=schema,
+      strategy='check',
+      unique_key='id',
+      check_cols=['order_date'],
+    )
+}}
+
+select * from {{ ref('orders') }}
+
+{% endsnapshot %}